logger_manager/pyproject.toml
[tool.poetry]
name = "logger_manager"
version = "0.1.0"
description = "to be done"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
    { include = "logger/manager", from = "src" },
]

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^2.0.0"
pandas = "^2.2.2"
python-dotenv = "^1.1.0"
jupytext = "^1.17.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.1"
pre-commit = "^3.7.0"
jupyter = "^1.0.0"
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
